Honeypot 1.2 This version contains some more checks for other types of Honeypot destruction methods such as Pistons and Silverfish. It also allows for colored config messages using & color codes, a customizable prefix, and container open actions. Finally, updated HoneypotUpdateChecker to use GitHub Raw instead of the Legacy Spigot API for version checking, as GitHub Raw is more reliable.
Added:
* Piston push/pull prevention on Honeypot blocks
* Silverfish infestation prevention (Needs testing, could not get silverfish to infest blocks at all, Honeypot and non-Honeypot)
* Ability to run Honeypot actions if containers are opened, not just broken

Changed:
* Switched to using GitHub instead of Spigot's legacy system for update checking. This is not noticeable on the end-user's side but makes development easier and update messages quicker to push

To-Do:
* Notify admins on server join of updates, not just on server start. Servers that rarely/never restart their server or pay attention to logs might miss update messages, so this will be a configurable option.

Known Issues:
* There is an issue which involves creating honeypot blocks with less-than-full blocks in front of them (For example, a lever). Imagine you are a player (P) creating a Honeypot block (H) with a lever in front (L). It would look like this P-L-H. Because the lever is in between you and the Honeypot block, it creates the Honeypot block on the lever, even if you're targeting the Honeypot block. I am using player.getTargetBlock() so I'm not sure why Spigot's API is causing this issue. I will investigate further for an alternative option, but for now, please ensure every block between you and the Honeypot block is air, and not something like a button, redstone dust, lever, flower, grass, etc.

+++ src/main/java/me/terrorbyte/honeypot/events/HoneypotEntityChangeEventListener.java
@@ -16,12 +16,17 @@ public class HoneypotEntityChangeEventListener implements Listener {
 
         //If the entity grabbing the block is an enderman, if they are allowed to, delete the Honeypot, otherwise cancel it
         if (event.getEntity().getType().equals(EntityType.ENDERMAN)){
-            if(Honeypot.getPlugin().getConfig().getBoolean("allow-enderman")){
-                HoneypotBlockStorageManager.deleteBlock(event.getBlock());
-            } else {
+            if(HoneypotBlockStorageManager.isHoneypotBlock(event.getBlock())) {
+                if (Honeypot.getPlugin().getConfig().getBoolean("allow-enderman")) {
+                    HoneypotBlockStorageManager.deleteBlock(event.getBlock());
+                } else {
+                    event.setCancelled(true);
+                }
+            }
+        } else if (event.getEntity().getType().equals(EntityType.SILVERFISH)) {
+            if(HoneypotBlockStorageManager.isHoneypotBlock(event.getBlock())){
                 event.setCancelled(true);
             }
         }
     }
-
 }
